+++ src/main/kotlin/lotto/Application.kt
@@ -1,5 +1,10 @@
 package lotto
 
+import lotto.controller.LottoGame
+import lotto.domain.LottoMachine
+import lotto.domain.Player
+import lotto.view.OutputView
+
 fun main() {
     val output = OutputView()
     val player = Player()
