+++ src/main/kotlin/lotto/domain/Player.kt
@@ -1,6 +1,7 @@
-package lotto
+package lotto.domain
 
 import camp.nextstep.edu.missionutils.Console
+import lotto.util.Constants
 
 
 class Player {
