+++ src/main/kotlin/lotto/domain/WinningResult.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class WinningResult(
     private val lottoTickets: List<Lotto>,
